Generation of measurement data

Set up

PastaQ.randombases — Function

randombases(N::Int, nbases::Int; local_basis = "Pauli")

Generate nbases measurement bases. By default, each local basis is randomly selected between ["X","Y","Z"], with "Z" being the default basis where the quantum state is written.

PastaQ.measurementgates — Function

measurementgates(basis::Vector)

Given as input a measurement basis, returns the corresponding gate data structure. If the basis is "Z", no action is required. If not, a quantum gate corresponding to the given basis rotation is added to the list.

Example: basis = ["X","Z","Z","Y"]

-> gate_list = [("basisX", 1),("basisY", 4)]

PastaQ.randompreparations — Function

randompreparations(N::Int, nshots::Int;
                   local_input_state = ["X+","X-","Y+","Y-","Z+","Z-"],
                   ndistinctstates = nothing)

Generate nshots input states to a quantum circuit. By default, each single-qubit state is randomly selected between the 6 eigenstates of the Pauli matrices, ["X+","X-","Y+","Y-","Z+","Z-"]. If ndistinctstates is provided, the output consist of numprep different input states, each being repeated nshots÷ndistinctstates times.

PastaQ.readouterror! — Function

readouterror!(measurement::Union{Vector, Matrix}, p1given0, p0given1)

Add readout error to a single projective measurement.

Arguments:

measurement: bit string of projective measurement outcome

p1given0: readout error probability 0 -> 1

p0given1: readout error probability 1 -> 0

Data generation

PastaQ.getsamples! — Function

getsamples!(M::Union{MPS,MPO};
            readout_errors = (p1given0 = nothing,
                              p0given1 = nothing))

Generate a single projective measurement in the MPS/MPO reference basis. If readout_errors is non-trivial, readout errors with given probabilities are applied to the measurement outcome.

PastaQ.getsamples — Function

getsamples(T::ITensor, nshots::Int)

Generate nshots projective measurements for an input quantum state T, which can either be a wavefunction or density matrix (dense).

PastaQ.getsamples!(M::Union{MPS,MPO}, nshots::Int; kwargs...)

Perform nshots projective measurements of a wavefunction |ψ⟩ or density operator ρ in the MPS/MPO reference basis. Each measurement consists of a binary vector σ = (σ₁,σ₂,…), drawn from the probabilty distribution:

P(σ) = |⟨σ|ψ⟩|² : if M = ψ is MPS

P(σ) = ⟨σ|ρ|σ⟩ : if M = ρ is MPO

getsamples(M::Union{MPS,MPO,ITensor}, bases::Matrix; kwargs...)

Generate a dataset of measurements acccording to a set of input bases, performing nshots measurements per basis. For a single measurement, Û is the depth-1 local circuit rotating each qubit, the data-point σ = (σ₁,σ₂,…) is drawn from the probability distribution:

P(σ) = |⟨σ|Û|ψ⟩|² : if M = ψ is MPS

P(σ) = <σ|Û ρ Û†|σ⟩ : if M = ρ is MPO