Quantum tomography

Tomography

PastaQ.tomography — Function

tomography(train_data::Matrix{Pair{String, Int}}, L::LPDO;
           optimizer::Optimizer,
           observer! = nothing,
           batchsize::Int64 = 100,
           epochs::Int64 = 1000,
           kwargs...)

Run quantum state tomography using a variational model L to fit train_data. The model can be either a pure state (MPS) or a density operator (LPDO).

Arguments:

train_data: pairs of basis/outcome: ("X"=>0, "Y"=>1, "Z"=>0, …).

L: variational model (MPS/LPDO).

optimizer: algorithm used to update the model parameters.

observer!: if provided, keep track of training metrics.

batch_size: number of samples used for one gradient update.

epochs: number of training iterations.

target: target quantum state (if provided, compute fidelities).

test_data: data for computing cross-validation.

outputpath: if provided, save metrics on file.

tomography(train_data::Matrix{Pair{String,Pair{String, Int}}}, L::LPDO;
           optimizer::Optimizer,
           observer! = nothing,
           batchsize::Int64 = 100,
           epochs::Int64 = 1000,
           kwargs...)

Run quantum process tomography using a variational model L to fit train_data. The model can be either a unitary circuit (MPO) or a Choi matrix (LPDO).

Arguments:

train_data: pairs of preparation/ (basis/outcome): ("X+"=>"X"=>0, "Z-"=>"Y"=>1, "Y+"=>"Z"=>0, …).

L: variational model (MPO/LPDO).

optimizer: algorithm used to update the model parameters.

observer!: if provided, keep track of training metrics.

batch_size: number of samples used for one gradient update.

epochs: number of training iterations.

target: target quantum process (if provided, compute fidelities).

test_data: data for computing cross-validation.

outputpath: if provided, save metrics on file.

PastaQ.split_dataset — Function

split_dataset(data::Matrix; train_ratio::Float64 = 0.9, randomize::Bool = true)

Split a data set into a train and test sets, given a train_ratio (i.e. the percentage of data in train_data. If randomize=true (default), the data is randomly shuffled before splitting.

Certification

PastaQ.fidelity — Function

fidelity(ρ::ITensor, σ::ITensor)

Compute the quantum fidelity between two ITensors, which are treated as density operators from the unprimed to the primed indices (if they are matrix-like).

Matrix-like ITensors should be Hermitian and non-negative.

Quantum state fidelity between two wavefunctions.

F = |⟨ψ|ϕ⟩|²

Quantum state fidelity between an MPS wavefunction and a density operator.

F = ⟨ψ|ρ|ψ⟩

Quantum state fidelity between an MPS wavefunction and a LPDO density operator.

F = ⟨ψ|ϱ|ψ⟩=|X†|ψ⟩|²

Quantum state/process fidelity between two MPO density matrices.

F = (tr[√(√A B √A)])²

If process = true, get process fidelities:

Quantum process fidelity between two MPO unitary operators. F = |⟨⟨A|B⟩⟩|²

Quantum process fidelity between a MPO unitary and MPO Choi matrix F = ⟨⟨A|B|A⟩⟩

Quantum process fidelity between two MPO Choi matrices F = (Tr[√(√A B √A)])²

Quantum process fidelity between a unitary MPO and a Choi LPDO. F = ⟨ψ|ϱ|ψ⟩=|X†|ψ⟩|²

Quantum process fidelity bewteen a Choi MPO and a Choi LPDO

Quantum fidelity between two LPDO density matrices.

F = (tr[√(√A B √A)])²

PastaQ.frobenius_distance — Function

frobenius_distance(ρ::Union{MPO, LPDO}, σ::Union{MPO, LPDO})

Compute the trace norm of the difference between two LPDOs and MPOs:

T(ρ,σ) = sqrt(trace[(ρ̃-σ̃)†(ρ̃-σ̃)])

where ρ̃ and σ̃ are the normalized density matrices.

PastaQ.fidelity_bound — Function

fidelity_bound(ρ::Union{MPO, LPDO}, σ::Union{MPO, LPDO})

Compute the the following lower bound of the fidelity:

F̃(ρ,σ) = trace[ρ̃† σ̃]

where ρ̃ and σ̃ are the normalized density matrices.

The bound becomes tight when the target state is nearly pure.