Circuit Simulator

Initialization

PastaQ.qubits — Function

qubits(N::Int; mixed::Bool=false)

qubits(sites::Vector{<:Index}; mixed::Bool=false)

Initialize qubits to:

An MPS wavefunction |ψ⟩ if mixed=false

An MPO density matrix ρ if mixed=true

PastaQ.resetqubits! — Function

resetqubits!(M::Union{MPS,MPO})

Reset qubits to the initial state:

|ψ⟩=|0,0,…,0⟩ if M = MPS

ρ = |0,0,…,0⟩⟨0,0,…,0| if M = MPO

Running a quantum circuit

PastaQ.runcircuit — Function

runcircuit(M::Union{MPS,MPO}, gate_tensors::Vector{<:ITensor}; kwargs...)

Apply the circuit to a state (wavefunction/densitymatrix) from a list of tensors.

runcircuit(M::Union{MPS,MPO}, gates::Vector{<:Tuple}; noise=nothing, apply_dag=nothing, 
           cutoff=1e-15, maxdim=10000)

Apply the circuit to a state (wavefunction or density matrix) from a list of gates.

If an MPS |ψ⟩ is input, there are three possible modes:

By default (noise = nothing and apply_dag = nothing), the evolution U|ψ⟩ is performed.

If noise is set to something nontrivial, the mixed evolution ε(|ψ⟩⟨ψ|) is performed. Example: noise = ("amplitude_damping", (γ = 0.1,)) (amplitude damping channel with decay rate γ = 0.1)

If noise = nothing and apply_dag = true, the evolution U|ψ⟩⟨ψ|U† is performed.

If an MPO ρ is input, there are three possible modes:

By default (noise = nothing and apply_dag = nothing), the evolution U ρ U† is performed.

If noise is set to something nontrivial, the evolution ε(ρ) is performed.

If noise = nothing and apply_dag = false, the evolution Uρ is performed.

runcircuit(N::Int, gates::Vector{<:Tuple}; process=false, noise=nothing,
           cutoff=1e-15, maxdim=10000, kwargs...)

Apply the circuit to a state (wavefunction or density matrix) from a list of gates, where the state has N physical qubits. The starting state is generated automatically based on the flags process, noise, and apply_dag.

By default (noise = nothing, apply_dag = nothing, and process = false), the evolution U|ψ⟩ is performed where the starting state is set to |ψ⟩ = |000...⟩. The MPS U|000...⟩ is returned.

If noise is set to something nontrivial, the mixed evolution ε(|ψ⟩⟨ψ|) is performed, where the starting state is set to |ψ⟩ = |000...⟩. The MPO ε(|000...⟩⟨000...|) is returned.

If process = true, the evolution U 1̂ is performed, where the starting state 1̂ = (1⊗1⊗1⊗…⊗1). The MPO approximation for the unitary represented by the set of gates is returned. In this case, noise must be nothing.

runcircuit(M::ITensor,gate_tensors::Vector{ <: ITensor}; kwargs...)

Apply the circuit to a ITensor from a list of tensors.

runcircuit(M::ITensor, gates::Vector{<:Tuple})

Apply the circuit to an ITensor from a list of gates.