Quantum tomography

Initialization

PastaQ.initializetomography — Function

initializetomography(N::Int64;kwargs...)

Initialize quantum tomography given a number of qubits

initializetomography(sites::Vector{<:Index}; kwargs...)

Initialize quantum tomography for a set of indices. By default, inintializes MPS, unless ξ is non-trivial, in which case it initializes a LPDO.

initializetomography(L::LPDO; kwargs...)

Initialize quantum tomography, given a reference state, either in MPS or MPO (LPDO) form

initializetomography(sites::Vector{<:Index},χ::Int64;σ::Float64=0.1)

Initialize a variational MPS for quantum tomography.

Arguments:

N: number of qubits

χ: bond dimension of the MPS

σ: width of initial box distribution

initializetomography(sites::Vector{<:Index},χ::Int64,ξ::Int64;
                     σ::Float64=0.1,purifier_tag = ts"Purifier")

Initialize a variational LPDO for quantum tomography.

Arguments:

N: number of qubits

χ: bond dimension of the LPDO

ξ: local purification dimension of the LPDO

σ: width of initial box distribution

Tomography

PastaQ.tomography — Function

tomography(L::LPDO, data::Array, opt::Optimizer; kwargs...)
tomography(ψ::MPS, data::Array, opt::Optimizer; kwargs...)

Run quantum state tomography using a the starting state model on data.

Arguments:

model: starting LPDO state.

data: training data set of projective measurements.

batchsize: number of data-points used to compute one gradient iteration.

epochs: total number of full sweeps over the dataset.

target: target quantum state underlying the data

choi: if true, compute probability using Choi matrix

observer: keep track of measurements and fidelities.

outputpath: write observer on file

tomography(L::LPDO,data_in::Array,data_out::Array,opt::Optimizer; kwargs...)
tomography(ψ::MPS,data_in::Array,data_out::Array,opt::Optimizer; kwargs...)

Run quantum process tomography on (data_in,data_out) using model as variational ansatz.

The data is reshuffled so it takes the format: (input1,output1,input2,output2,…).

Certification

PastaQ.fidelity — Function

fidelity(ψ::MPS,ϕ::MPS)

Compute the fidelity between two MPS:

F = |⟨ψ|ϕ⟩|²

fidelity(ψ::MPS, ρ::LPDO)

fidelity(ρ::LPDO, ψ::MPS)

Compute the fidelity between an MPS and LPDO.

F = ⟨ψ|ρ|ψ⟩

fidelity(ψ::MPS,ρ::MPO)

fidelity(ρ::MPO,ψ::MPS)

Compute the fidelity between an MPS and MPO.

F = ⟨ψ|ρ|ψ⟩

PastaQ.frobenius_distance — Function

frobenius_distance(ρ0::LPDO, σ0::LPDO)
frobenius_distance(ρ0::LPDO, σ0::MPO)
frobenius_distance(ρ0::MPO,  σ0::LPDO)
frobenius_distance(ρ0::MPO,  σ0::MPO)

Compute the trace norm of the difference between two LPDOs and MPOs.

T(ρ,σ) = sqrt(trace[(ρ-σ)†(ρ-σ)])