Quantum tomography

Initialization

Missing docstring.

Missing docstring for initializetomography. Check Documenter's build log for details.

Tomography

PastaQ.tomography — Function

tomography(L::LPDO,data::Array,opt::Optimizer; kwargs...)

TEMPORARY WRAPPER This function is required as long as the process tomography is using a split representation.

tomography(L::LPDO,data_in::Array,data_out::Array,opt::Optimizer; kwargs...)

TEMPORARY WRAPPER FOR UNSPLIT PROCESS TOMOGRAPHY

This function take a model L and a target (is provided) in a unsplit representation, and run tomography with the split algorithm. Returns the unsplit result.

Certification

PastaQ.fidelity — Function

fidelity(ψ::MPS, ϕ::MPS)

Compute the fidelity between two MPS:

F = |⟨ψ̃|ϕ̃⟩|²

where ψ̃ and ϕ̃ are the normalized MPS.

fidelity(ψ::MPS, ρ::Union{MPO, LPDO})
fidelity(ρ::Union{MPO, LPDO}, ψ::MPS)

Compute the fidelity between an MPS and MPO/LPDO:

F = ⟨ψ̃|ρ̃|ψ̃⟩

where ψ̃ and ρ̃ are the normalized MPS and MPO/LDPO.

PastaQ.frobenius_distance — Function

frobenius_distance(ρ::Union{MPO, LPDO}, σ::Union{MPO, LPDO})

Compute the trace norm of the difference between two LPDOs and MPOs:

T(ρ,σ) = sqrt(trace[(ρ̃-σ̃)†(ρ̃-σ̃)])

where ρ̃ and σ̃ are the trace normalized density matrices.