Quantum tomography

Tomography

PastaQ.tomography — Function

tomography(train_data::Matrix{Pair{String,Pair{String, Int}}}, L::LPDO;
           optimizer::Optimizer,
           observer! = nothing,
           batchsize::Int64 = 100,
           epochs::Int64 = 1000,
           kwargs...)

Run quantum process tomography using a variational model L to fit train_data. The model can be either a unitary circuit (MPO) or a Choi matrix (LPDO).

Arguments:

train_data: pairs of preparation/ (basis/outcome): ("X+"=>"X"=>0, "Z-"=>"Y"=>1, "Y+"=>"Z"=>0, …).

L: variational model (MPO/LPDO).

optimizer: algorithm used to update the model parameters.

observer!: if provided, keep track of training metrics.

batch_size: number of samples used for one gradient update.

epochs: number of training iterations.

target: target quantum process (if provided, compute fidelities).

test_data: data for computing cross-validation.

outputpath: if provided, save metrics on file.

PastaQ.split_dataset — Function

split_dataset(data::Matrix; train_ratio::Float64 = 0.9, randomize::Bool = true)

Split a data set into a train and test sets, given a train_ratio (i.e. the percentage of data in train_data. If randomize=true (default), the data is randomly shuffled before splitting.

Certification

PastaQ.fidelity — Function

fidelity(ψ::MPS, ϕ::MPS; kwargs...)

Quantum state fidelity between two wavefunctions.

F = |⟨ψ|ϕ⟩|²

fidelity(ψ::MPS, ρ::MPO; kwargs...)
fidelity(ρ::MPO, ψ::MPS; kwargs...)

Quantum state fidelity between an MPS wavefunction and a density operator.

F = ⟨ψ|ρ|ψ⟩

fidelity(Ψ::MPS, ϱ::LPDO{MPO}; cutoff::Float64 = 1e-15)
fidelity(ϱ::LPDO{MPO}, ψ::MPS; kwargs...)

Quantum state fidelity between an MPS wavefunction and a LPDO density operator.

F = ⟨ψ|ϱ|ψ⟩=|X†|ψ⟩|²

fidelity(A::MPO, B::MPO; process::Bool = false, cutoff::Float64 = 1e-15)

Fidelity between two MPOs. If any is a Choi matrix, wrap them with the Choi type.

fidelity(A::MPO, B::LPDO{MPO}; process::Bool=false, cutoff::Float64 = 1e-15)

Fidelity between a MPO and a LPDO. Wrap the MPO in the Choi if so.

fidelity(A::LPDO{MPO}, B::LPDO{MPO}; kwargs...)

Quantum fidelity between two LPDO density matrices.

fidelity(A::ITensor, B::ITensor)

Compute the quantum fidelity between two ITensors. Wrap each one in the ITensorState and the ITensorOperator according to the index structure to allow dispatch.

fidelity(A::ITensorState, B::ITensorState; kwargs...)

State fidelity between two wavefunctions

fidelity(A::ITensorState, B::ITensorOperator; kwargs...)
fidelity(A::ITensorOperator, B::ITensorState; kwargs...)

State fidelity between a wavefunction and a density operator

fidelity(A::ITensorOperator, B::ITensorOperator; kwargs...)

Fidelity between two operators. Wrap the Choi type as for the TN (above)

PastaQ.frobenius_distance — Function

frobenius_distance(ρ::Union{MPO, LPDO}, σ::Union{MPO, LPDO})

Compute the trace norm of the difference between two LPDOs and MPOs:

T(ρ,σ) = sqrt(trace[(ρ̃-σ̃)†(ρ̃-σ̃)])

where ρ̃ and σ̃ are the normalized density matrices.

PastaQ.fidelity_bound — Function

fidelity_bound(ρ::Union{MPO, LPDO}, σ::Union{MPO, LPDO})

Compute the the following lower bound of the fidelity:

F̃(ρ,σ) = trace[ρ̃† σ̃]

where ρ̃ and σ̃ are the normalized density matrices.

The bound becomes tight when the target state is nearly pure.